SRA-008 implement Rule Processor and its unit tests and refactored code in ReplyMessageProcessor
